kann jetzt soviele Seifenblasen erstellen wie ich will, sie fallen gerade oder steigen gerade nachdem ich sie beruehrt habe, kann beliebig viele punkte nachpruefen ob sie die seifenblasen beruehren, jetzt muss man sich ueberlegen wie man das mit linien und nicht nur mit punkten macht ausserdem waere es schoen wenn man nicht SOIL brauchte um die textur zu laden

+++ Blase.h
@@ -17,11 +17,11 @@ public:
 	void setMove( int x, int y);
 
 	int r;	//radius
-	int x;	//X Koordinate
-	int y;	//Y Koordinate
+	double x;	//X Koordinate
+	double y;	//Y Koordinate
 
-	int vx;	// Bewegungsrichtung der x Koordinate mit Geschwindigkeit
-	int vy;  // Bewegungsrichtung der y Koordinate mit Geschwindigkeit
+	double vx;	// Bewegungsrichtung der x Koordinate mit Geschwindigkeit
+	double vy;  // Bewegungsrichtung der y Koordinate mit Geschwindigkeit
 
 
 };
